Move search wallet and tooltip SCSS to CSS
