--- components/delegation/walletChecker/WalletChecker.module.scss
@@ -1,154 +1,0 @@
-@use "sass:color";
-@use "../../../styles/variables.scss" as variables;
-
-.formInput {
-  color: variables.$dark-grey !important;
-  border-radius: 0 !important;
-  border-color: variables.$lightest-grey;
-
-  &:focus {
-    border-color: variables.$project-blue !important;
-  }
-}
-
-.intro,
-.statusText {
-  color: variables.$font-color-h;
-  margin-bottom: 0.75rem;
-}
-
-.checkBtn {
-  background-color: variables.$project-blue;
-  border: none;
-  border-radius: 0;
-  color: white;
-  padding: 10px 15px;
-
-  &:hover {
-    background-color: color.adjust(
-      $color: variables.$project-blue,
-      $lightness: 5%
-    );
-  }
-
-  &:active,
-  &:disabled {
-    background-color: color.scale(
-      variables.$project-blue,
-      $lightness: -5%
-    ) !important;
-  }
-}
-
-.clearBtn {
-  background-color: transparent;
-  border-radius: 0;
-  color: white;
-  padding: 10px 15px;
-
-  &:hover {
-    background-color: variables.$dark-grey;
-  }
-
-  &:active,
-  &:disabled {
-    background-color: variables.$dark-grey;
-  }
-}
-
-.error {
-  font-size: 1em;
-  color: variables.$raspberry;
-}
-
-.supportingAddress {
-  a {
-    font-weight: normal;
-  }
-}
-
-.consolidationPlusIcon {
-  height: 20px;
-  margin-left: 10px;
-  margin-right: 15px;
-}
-
-.arrowHead {
-  display: inline-block;
-  width: 0;
-  height: 0;
-  border-top: 8px solid transparent;
-  border-bottom: 8px solid transparent;
-  border-left: 8px solid white; /* Customize the color and style as needed */
-  margin-right: 25px;
-}
-
-.arrowBody {
-  display: inline-block;
-  margin-left: 20px;
-  width: 75px;
-  height: 4px;
-  background-color: white;
-}
-
-.arrowBody::after {
-  content: "";
-  position: absolute;
-  top: 50%;
-  left: 100%;
-  transform: translate(-50%, -50%);
-  width: 50px; /* Adjust the length of the line */
-  height: 2px; /* Adjust the thickness of the line */
-}
-
-.recommendationsList {
-  padding-left: 1rem;
-}
-
-.consolidationRecommendationIcon {
-  height: 30px;
-  color: red;
-  margin-right: 10px;
-}
-
-.activeDelegationIcon {
-  width: 30px;
-  height: 30px;
-  color: white;
-  background-color: green;
-  padding: 5px;
-  border-radius: 18px;
-}
-
-.consolidationActiveIcon {
-  width: 30px;
-  height: 30px;
-  color: white;
-  margin-left: 25px;
-  background-color: green;
-  padding: 5px;
-  border-radius: 18px;
-}
-
-.walletCheckerContainer {
-  max-height: 450px;
-  overflow-y: auto;
-}
-
-.walletOk {
-  &:focus {
-    border-color: variables.$project-blue !important;
-  }
-}
-
-.walletInput {
-  all: unset;
-  width: 100%;
-  padding: 10px;
-  border: 1px solid variables.$lighter-grey;
-  border-radius: 18px;
-}
-
-.infoOk {
-  color: greenyellow;
-}
