--- components/searchModal/SearchModal.module.scss
@@ -1,208 +1,0 @@
-@use "sass:color";
-@use "../../styles/variables.scss";
-
-.modalInput {
-  border-radius: 0;
-  border: 1px solid variables.$very-dark-grey;
-  color: variables.$very-dark-grey !important;
-
-  &:focus {
-    border-color: black;
-  }
-}
-
-.removeWalletBtn {
-  appearance: none;
-  background: transparent;
-  border: 0;
-  color: inherit;
-  cursor: pointer;
-  display: inline-flex;
-  padding: 0;
-  vertical-align: middle;
-  width: 15px;
-}
-
-.shakeWalletInput {
-  input {
-    border-color: variables.$raspberry !important;
-  }
-  button {
-    background-color: variables.$raspberry !important;
-  }
-  /* Start the shake animation and make the animation last for 0.5 seconds */
-  animation: shake 0.2s;
-
-  /* When the animation is finished, start again */
-  animation-iteration-count: infinite;
-}
-
-@keyframes shake {
-  0% {
-    transform: translate(1px, 1px) rotate(0deg);
-  }
-  10% {
-    transform: translate(-1px, -2px) rotate(-1deg);
-  }
-  20% {
-    transform: translate(-3px, 0px) rotate(1deg);
-  }
-  30% {
-    transform: translate(3px, 2px) rotate(0deg);
-  }
-  40% {
-    transform: translate(1px, -1px) rotate(1deg);
-  }
-  50% {
-    transform: translate(-1px, 2px) rotate(-1deg);
-  }
-  60% {
-    transform: translate(-3px, 1px) rotate(0deg);
-  }
-  70% {
-    transform: translate(3px, 1px) rotate(-1deg);
-  }
-  80% {
-    transform: translate(-1px, -1px) rotate(1deg);
-  }
-  90% {
-    transform: translate(1px, 2px) rotate(0deg);
-  }
-  100% {
-    transform: translate(1px, -2px) rotate(-1deg);
-  }
-}
-
-.modalButton {
-  background-color: variables.$very-dark-grey;
-  border: none !important;
-  border-radius: 0;
-
-  &:hover,
-  &:active {
-    background-color: black !important;
-  }
-}
-
-.modalButtonClear {
-  background-color: transparent;
-  border-color: variables.$dark-grey;
-  color: variables.$dark-grey !important;
-  border-radius: 0;
-
-  &:hover {
-    background-color: transparent !important;
-    color: black !important;
-    border-color: black !important;
-  }
-
-  &:active {
-    background-color: variables.$lightest-grey !important;
-    border-color: variables.$lightest-grey !important;
-  }
-
-  &:disabled {
-    background-color: variables.$lightest-grey !important;
-    border-color: variables.$lightest-grey !important;
-  }
-}
-
-.modalButtonDone {
-  float: right;
-  border-color: variables.$very-dark-grey;
-  background-color: variables.$very-dark-grey;
-  color: variables.$font-color;
-  border-radius: 0;
-
-  &:hover {
-    background-color: variables.$dark-grey;
-    border-color: variables.$dark-grey;
-  }
-}
-
-.noSearchWalletsText {
-  color: variables.$font-color-h;
-}
-
-.searchBtn {
-  background-color: white !important;
-  color: black;
-  width: 40px;
-  height: 40px;
-  text-align: center;
-  border-radius: 20px;
-  padding: 10px;
-  cursor: pointer;
-  &:hover {
-    background-color: variables.$off-white !important;
-  }
-
-  &:disabled {
-    background-color: variables.$lightest-grey !important;
-    cursor: not-allowed;
-  }
-}
-
-.searchBtnActive {
-  background-color: color.adjust(
-    $color: variables.$project-green,
-    $lightness: 10%
-  ) !important;
-
-  &:hover,
-  &:active,
-  &:focus {
-    background-color: variables.$project-green !important;
-  }
-}
-
-.searchWalletDisplay {
-  background-color: #333;
-  padding: 5px 10px;
-  margin-right: 10px;
-  border-top-right-radius: 5px;
-  border-bottom-right-radius: 5px;
-}
-
-.searchWalletDisplayBtn {
-  width: 30px;
-  background-color: #333 !important;
-  padding: 3px 5px !important;
-  border-right: 1px solid #999 !important;
-  border-top-left-radius: 5px;
-  border-bottom-left-radius: 5px;
-  cursor: pointer;
-  outline: 0 !important;
-  text-decoration: none !important;
-
-  &:hover {
-    background-color: variables.$lighter-grey !important;
-  }
-
-  &:disabled {
-    background-color: variables.$lightest-grey !important;
-    cursor: not-allowed;
-  }
-}
-
-.clearSearchBtnIcon {
-  align-items: center;
-  appearance: none;
-  background: transparent;
-  border: 0;
-  color: inherit;
-  display: inline-flex;
-  justify-content: center;
-  width: 40px;
-  height: 40px;
-  margin-right: 10px;
-  cursor: pointer;
-
-  &:hover {
-    color: variables.$lightest-grey;
-  }
-
-  &:focus {
-    outline: none !important;
-  }
-}
