+++ components/utils/tooltip/CustomTooltip.module.css
@@ -1,0 +1,279 @@
+.tooltip {
+  position: fixed;
+  z-index: 999999;
+  visibility: hidden;
+  opacity: 0;
+  animation: tooltipIn 0.15s cubic-bezier(0.4, 0, 0.2, 1) forwards;
+  transition:
+    opacity 0.15s ease,
+    visibility 0.15s ease,
+    transform 0.15s cubic-bezier(0.4, 0, 0.2, 1);
+}
+
+.tooltip--top {
+  transform-origin: center bottom;
+}
+
+.tooltip--bottom,
+.tooltip--auto {
+  transform-origin: center top;
+}
+
+.tooltip--left {
+  transform-origin: right center;
+}
+
+.tooltip--right {
+  transform-origin: left center;
+}
+
+@keyframes tooltipIn {
+  0% {
+    visibility: hidden;
+    transform: scale(0.95);
+    opacity: 0;
+  }
+
+  1% {
+    visibility: hidden;
+    transform: scale(0.95);
+    opacity: 0;
+  }
+
+  100% {
+    visibility: visible;
+    transform: scale(1);
+    opacity: 1;
+  }
+}
+
+.tooltipContent {
+  position: relative;
+  overflow: hidden;
+  border: 1px solid rgba(255, 255, 255, 0.08);
+  border-radius: 12px;
+  background: #131316;
+  padding: 12px 16px;
+  color: #fff;
+  font-family:
+    "Montserrat",
+    -apple-system,
+    BlinkMacSystemFont,
+    "Segoe UI",
+    sans-serif;
+  font-size: 14px;
+  font-weight: 400;
+  line-height: 1.5;
+  text-rendering: optimizeLegibility;
+  box-shadow:
+    0 24px 48px rgba(0, 0, 0, 0.68),
+    inset 0 1px 0 rgba(255, 255, 255, 0.03);
+  user-select: none;
+  -webkit-font-smoothing: antialiased;
+  -moz-osx-font-smoothing: grayscale;
+  -webkit-user-select: none;
+}
+
+.tooltipContent a:not([class]) {
+  color: #4a9eff;
+  font-weight: 500;
+  text-decoration: none;
+}
+
+.tooltipContent a:not([class]):hover {
+  text-decoration: underline;
+  text-underline-offset: 2px;
+}
+
+.tooltipContent :global(.tailwind-scope) :global(.tw-font-bold) {
+  font-weight: 700;
+}
+
+.tooltipArrow {
+  pointer-events: none;
+  position: absolute;
+  width: 0;
+  height: 0;
+}
+
+.tooltipArrow--top {
+  bottom: -6px;
+  border-top: 6px solid rgba(255, 255, 255, 0.08);
+  border-right: 6px solid transparent;
+  border-left: 6px solid transparent;
+}
+
+.tooltipArrow--top::after {
+  position: absolute;
+  bottom: 1px;
+  left: -5px;
+  width: 0;
+  height: 0;
+  border-top: 5px solid #131316;
+  border-right: 5px solid transparent;
+  border-left: 5px solid transparent;
+  content: "";
+}
+
+.tooltipArrow--bottom {
+  top: -6px;
+  border-right: 6px solid transparent;
+  border-bottom: 6px solid rgba(255, 255, 255, 0.08);
+  border-left: 6px solid transparent;
+}
+
+.tooltipArrow--bottom::after {
+  position: absolute;
+  top: 1px;
+  left: -5px;
+  width: 0;
+  height: 0;
+  border-right: 5px solid transparent;
+  border-bottom: 5px solid #131316;
+  border-left: 5px solid transparent;
+  content: "";
+}
+
+.tooltipArrow--left {
+  right: -6px;
+  border-top: 6px solid transparent;
+  border-bottom: 6px solid transparent;
+  border-left: 6px solid rgba(255, 255, 255, 0.08);
+}
+
+.tooltipArrow--left::after {
+  position: absolute;
+  top: -5px;
+  left: -5px;
+  width: 0;
+  height: 0;
+  border-top: 5px solid transparent;
+  border-bottom: 5px solid transparent;
+  border-left: 5px solid #131316;
+  content: "";
+}
+
+.tooltipArrow--right {
+  left: -6px;
+  border-top: 6px solid transparent;
+  border-right: 6px solid rgba(255, 255, 255, 0.08);
+  border-bottom: 6px solid transparent;
+}
+
+.tooltipArrow--right::after {
+  position: absolute;
+  top: -5px;
+  right: -5px;
+  width: 0;
+  height: 0;
+  border-top: 5px solid transparent;
+  border-right: 5px solid #131316;
+  border-bottom: 5px solid transparent;
+  content: "";
+}
+
+@media (max-width: 768px) {
+  .tooltipContent {
+    padding: 10px 14px;
+    font-size: 13px;
+  }
+
+  .tooltipArrow--top {
+    bottom: -5px;
+    border-top-width: 5px;
+    border-right-width: 5px;
+    border-left-width: 5px;
+  }
+
+  .tooltipArrow--top::after {
+    left: -4px;
+    border-top-width: 4px;
+    border-right-width: 4px;
+    border-left-width: 4px;
+  }
+
+  .tooltipArrow--bottom {
+    top: -5px;
+    border-right-width: 5px;
+    border-bottom-width: 5px;
+    border-left-width: 5px;
+  }
+
+  .tooltipArrow--bottom::after {
+    left: -4px;
+    border-right-width: 4px;
+    border-bottom-width: 4px;
+    border-left-width: 4px;
+  }
+
+  .tooltipArrow--left {
+    right: -5px;
+    border-top-width: 5px;
+    border-bottom-width: 5px;
+    border-left-width: 5px;
+  }
+
+  .tooltipArrow--left::after {
+    top: -4px;
+    left: -4px;
+    border-top-width: 4px;
+    border-bottom-width: 4px;
+    border-left-width: 4px;
+  }
+
+  .tooltipArrow--right {
+    left: -5px;
+    border-top-width: 5px;
+    border-right-width: 5px;
+    border-bottom-width: 5px;
+  }
+
+  .tooltipArrow--right::after {
+    top: -4px;
+    right: -4px;
+    border-top-width: 4px;
+    border-right-width: 4px;
+    border-bottom-width: 4px;
+  }
+}
+
+@media (prefers-contrast: high) {
+  .tooltipContent {
+    border-width: 2px;
+    border-color: #fff;
+    background: #000;
+    color: #fff;
+  }
+
+  .tooltipArrow--top {
+    border-top-color: #fff;
+  }
+
+  .tooltipArrow--top::after {
+    border-top-color: #000;
+  }
+
+  .tooltipArrow--bottom {
+    border-bottom-color: #fff;
+  }
+
+  .tooltipArrow--bottom::after {
+    border-bottom-color: #000;
+  }
+
+  .tooltipArrow--left {
+    border-left-color: #fff;
+  }
+
+  .tooltipArrow--left::after {
+    border-left-color: #000;
+  }
+
+  .tooltipArrow--right {
+    border-right-color: #fff;
+  }
+
+  .tooltipArrow--right::after {
+    border-right-color: #000;
+  }
+}
